implement images in progress

+++ src/main/java/com/example/petfinder/service/impl/AnimalServiceImpl.java
@@ -2,8 +2,8 @@ package com.example.petfinder.service.impl;
 
 import com.example.petfinder.dto.animal.request.AnimalCreation;
 import com.example.petfinder.dto.animal.request.AnimalUpdating;
-import respose.AnimalCard;
-import respose.AnimalView;
+import com.example.petfinder.dto.animal.respose.AnimalCard;
+import com.example.petfinder.dto.animal.respose.AnimalView;
 import com.example.petfinder.exceptions.EntityNotExistsException;
 import com.example.petfinder.exceptions.PermissionException;
 import com.example.petfinder.mapper.AnimalMapper;
@@ -11,13 +11,17 @@ import com.example.petfinder.mapper.ImageMapper;
 import com.example.petfinder.model.entity.Animal;
 import com.example.petfinder.model.entity.Image;
 import com.example.petfinder.model.entity.User;
+import com.example.petfinder.model.enums.Status;
 import com.example.petfinder.repository.AnimalRepository;
 import com.example.petfinder.repository.ImageRepository;
 import com.example.petfinder.repository.UserRepository;
 import com.example.petfinder.service.AnimalService;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 
 import java.io.IOException;
@@ -26,6 +30,7 @@ import java.util.zip.DataFormatException;
 
 @Service
 @RequiredArgsConstructor
+@Slf4j
 public class AnimalServiceImpl implements AnimalService {
     private final AnimalRepository animalRepository;
     private final UserRepository userRepository;
@@ -42,10 +47,12 @@ public class AnimalServiceImpl implements AnimalService {
 
 
     @Override
-    public AnimalView addAnimal(AnimalCreation animalCreation, String ownerEmail) throws IOException, DataFormatException {
+    public AnimalView addAnimal(AnimalCreation animalCreation) throws IOException, DataFormatException {
+        User user = getCurrentUser();
         Animal animal = animalMapper.newToAnimal(animalCreation);
-        User user = findUserByLogin(ownerEmail);
+
         animal.setOwner(user);
+        animal.setStatus(Status.NEED_ADOPTION);
         animal = animalRepository.save(animal);
         return animalMapper.animalToView(animal);
     }
@@ -61,12 +68,10 @@ public class AnimalServiceImpl implements AnimalService {
     }
 
     @Override
-    public AnimalView updateAnimal(UUID animalId, AnimalUpdating animalUpdate, String ownerEmail) throws IOException, DataFormatException {
+    public AnimalView updateAnimal(UUID animalId, AnimalUpdating animalUpdate) throws IOException, DataFormatException {
+        checkPermission(animalId);
+
         Animal animal = findAnimalById(animalId);
-        User user = findUserByLogin(ownerEmail);
-        if (!user.getEmail().equals(animal.getOwner().getEmail())) {
-            throw new PermissionException(PERMISSION_DENIED);
-        }
         copyUpdateFieldsToAnimal(animalUpdate, animal);
         animal = animalRepository.save(animal);
         return animalMapper.animalToView(animal);
@@ -74,6 +79,7 @@ public class AnimalServiceImpl implements AnimalService {
 
     @Override
     public void deleteAnimal(UUID animalId) {
+        checkPermission(animalId);
         animalRepository.deleteById(animalId);
 
     }
@@ -82,9 +88,6 @@ public class AnimalServiceImpl implements AnimalService {
         return animalRepository.findById(animalId).orElseThrow(() -> new EntityNotExistsException(String.format(ANIMAL_NOT_FOUND, animalId)));
     }
 
-    private User findUserByLogin(String ownerLogin) {
-        return userRepository.findUserByEmail(ownerLogin).orElseThrow(() -> new EntityNotExistsException(String.format(USER_WITH_LOGIN_NOT_FOUND, ownerLogin)));
-    }
 
     private void copyUpdateFieldsToAnimal(AnimalUpdating animalUpdating, Animal animal) throws IOException {
         Image image = (animalUpdating.image() != null) ? imageMapper.multiPartFileToImage(animalUpdating.image()) : null;
@@ -95,11 +98,34 @@ public class AnimalServiceImpl implements AnimalService {
         animal.setName(animalUpdating.name());
         animal.setType(animalUpdating.type());
         animal.setWeight(animalUpdating.weight());
-        animal.setBirthDate(animalUpdating.birthDate());
-        animal.setImage(image);
+        animal.setBirthday(animalUpdating.birthday());
+        animal.setImage1(image);
+        animal.setImage2(image);
+        animal.setImage3(image);
+        animal.setImage4(image);
         animal.setSterilization(animalUpdating.sterilization());
-        animal.setDescription(animalUpdating.description());
+        animal.setStory(animalUpdating.story());
+        animal.setBreed(animalUpdating.breed());
+        animal.setTraits(animalUpdating.traits());
+        animal.setHealthHistory(animalUpdating.healthHistory());
         animal.setSize(animalUpdating.size());
+    }
+
+
+    public User getByUsername(String username) {
+        return userRepository.findUserByEmail(username).orElseThrow(() -> new UsernameNotFoundException("User not found"));
+
+    }
+
+    public User getCurrentUser() {
+        var username = SecurityContextHolder.getContext().getAuthentication().getName();
+        return getByUsername(username);
+    }
+
+    private void checkPermission(UUID animalId) {
+        if (!getCurrentUser().getAnimals().contains(findAnimalById(animalId))) {
+            throw new PermissionException();
+        }
     }
 
 }
